# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inngest App API",
    "version": "1.0.0",
    "description": "API documentation for the Inngest app"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format: Bearer {token}"
      }
    },
    "schemas": {
...
      "WorkflowInputField": {
        "type": "object",
        "required": [
          "key",
          "label",
          "type"
        ],
        "properties": {
          "key": {
            "type": "string",
            "minLength": 1
          },
          "label": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "boolean",
              "json",
              "credential",
              "email",
              "url",
              "date",
              "time",
              "file",
              "select",
              "multiselect"
            ]
          },
          "required": {
            "type": "boolean",
            "default": false
          },
          "defaultValue": {},
          "validation": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              },
              "pattern": {
                "type": "string"
              },
              "options": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "WorkflowTemplateResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "category",
          "eventName"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "REPORTING",
              "DATA_PROCESSING",
              "COMMUNICATION",
              "AUTOMATION",
              "ANALYTICS",
              "INTEGRATION",
              "MONITORING",
              "CUSTOM"
            ]
          },
          "version": {
            "type": "string",
            "default": "1.0.0"
          },
          "eventName": {
            "type": "string"
          },
          "canBeScheduled": {
            "type": "boolean",
            "default": false
          },
          "requiredProviders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "SLACK",
                "HUBSPOT",
                "FIRECRAWL",
                "CUSTOM"
              ]
            }
          },
          "inputFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowInputField"
            }
          },
          "outputSchema": {
            "type": "object"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/google/callback": {
      "get": {
        "summary": "Google OAuth callback",
        "description": "Handles the Google OAuth callback.",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to the integrations page."
          },
          "400": {
            "description": "Missing code or state from Google OAuth callback."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
    "/api/marketplace/workflows": {
      "get": {
        "summary": "Get available workflow templates",
        "description": "Returns workflow templates available for installation. Includes default templates (auto-installed) and marketplace templates that can be added.",
        "tags": [
          "Marketplace"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "REPORTING",
                "DATA_PROCESSING",
                "COMMUNICATION",
                "AUTOMATION",
                "ANALYTICS",
                "INTEGRATION",
                "MONITORING",
                "CUSTOM"
              ]
            },
            "description": "Filter by workflow category"
          },
          {
            "in": "query",
            "name": "provider",
            "schema": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "SLACK",
                "HUBSPOT",
                "FIRECRAWL",
                "CUSTOM"
              ]
            },
            "description": "Filter by required provider"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "default",
                "marketplace",
                "custom"
              ]
            },
            "description": "Filter by template type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of available workflow templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WorkflowTemplateResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}